+++ src/main/java/dooya/see/domain/post/event/PostPublished.java
@@ -1,4 +1,4 @@
-package dooya.see.domain.post;
+package dooya.see.domain.post.event;
 
 import dooya.see.domain.shared.DomainEvent;
