+++ src/test/java/dooya/see/domain/post/PostTest.java
@@ -1,9 +1,6 @@
 package dooya.see.domain.post;
 
-import dooya.see.domain.post.event.PostCreated;
-import dooya.see.domain.post.event.PostDeleted;
-import dooya.see.domain.post.event.PostHidden;
-import dooya.see.domain.post.event.PostUpdated;
+import dooya.see.domain.post.event.*;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
